Decouple contact account association from VCardEntry construction
Bluetooth receives and parses VCardEntry objects from a remote device.
The objects we receive have no concept of an Android AccountManager
framework Account object. We would like to be able to add that account,
just in time, before creating the insertion operations for Contacts
Provider. This way, the download and parsing mechanism doesn't need to
carry implementation specifics of the insertion mechanism.

Bug: 367323870
Test: atest com.android.vcard.tests.VCardEntryTests
Change-Id: Ie2a262e75bd218ccd883434028df86c9a438a400